requirements.txt
numpy>=1.26
pandas>=2.1
scipy>=1.11
statsmodels>=0.14
fredapi>=0.5
pyyaml>=6.0
python-dotenv>=1.0
plotly>=5.17
streamlit>=1.28
matplotlib>=3.8
seaborn>=0.13
scikit-learn>=1.3
langchain>=0.1
langchain-openai>=0.0.5
beautifulsoup4>=4.12
requests>=2.31
duckdb>=0.9
loguru>=0.7
pytest>=7.4
pytest-cov>=4.1
ruff>=0.1
black>=23.0
